+++ src/datelogic/simpledate.js
@@ -1,4 +1,6 @@
-const dateRegex = /(\d{2})\.(\d{2})\.(\d{4})/;
+import errorTypes from './error-types';
+
+const dateRegex = /^(\d{1,2})(\.| |-|\/)?(\d{1,2})\2(\d{2}(\d{2})?)$/;
 
 function isDate(date) {
   return !!date && typeof date === 'object' &&
@@ -6,9 +8,7 @@ function isDate(date) {
 }
 
 function SimpleDate(date) {
-  if (typeof date === 'string') {
-    this.internalDate = new Date(date.replace(dateRegex, '$3-$2-$1'));
-  } else if (isDate(date)) {
+  if (isDate(date)) {
     this.internalDate = date;
   } else {
     this.internalDate = new Date();
@@ -24,25 +24,56 @@ SimpleDateFactory.today = function today() {
   return new SimpleDate();
 };
 
-SimpleDateFactory.fromString = function fromString(string) {
+function beforeYear(year, yearLimit) {
+  return parseInt(year, 10) < yearLimit;
+}
+
+/*
+  If given a two-digit year we assume this century.
+  E.g: 01.01.17 will yield 01.01.2017, not 01.01.1917 (JS default)
+  Else they'll have to use 4-digit year.
+*/
+function deriveTwoDigitYear(year) {
+  const yearDate = new Date(year, 0);
+  const fullYear = yearDate.getFullYear();
+  if (beforeYear(fullYear, 2000)) {
+    return yearDate.getFullYear() + 100;
+  }
+
+  return yearDate.getFullYear();
+}
+
+SimpleDateFactory.fromString = function fromString(string, onSuccess, onError) {
+  const success = onSuccess || (() => {});
+  const error = onError || (() => {});
   const newDate = new SimpleDate();
   const match = dateRegex.exec(string);
 
   if (!match) {
+    error(errorTypes.INVALID_DATE_FORMAT);
     return null;
   }
 
-  const [, date, month, year] = match;
+  const [ , date, , month] = match;
+  let year = match[4];
+
+  if (year.length === 2) {
+    year = deriveTwoDigitYear(year);
+  }
 
   newDate.setYear(year);
   newDate.setMonth(month - 1); // Months are 0-indexed
   newDate.setDate(date); // Need to set date last, to avoid month-overflow
-  if (newDate.date() !== date * 1 ||
+  if (
+    newDate.date() !== date * 1 ||
     newDate.month() + 1 !== month * 1 ||
-    newDate.year() !== year * 1) {
+    newDate.year() !== year * 1
+  ) {
+    error(errorTypes.INVALID_DATE);
     return null;
   }
 
+  success(newDate);
   return newDate;
 };