+++ src/i18n/nb.js
@@ -42,4 +42,8 @@ export default {
   PREVIOUS_SHORT: 'Forrige',
   NEXT_SHORT: 'Neste',
   FIRST_DAY_OF_WEEK: 1,
+  INVALID_DATE_FORMAT: 'Dato må skrives på dette formatet: dd.mm.åååå',
+  INVALID_DATE: 'Ugyldig dato',
+  MIN_DATE: 'Oppgitt dato er for langt tilbake',
+  MAX_DATE: 'Oppgitt date er for lang frem'
 };